Create README.md
This R program is written for calculating the pLDDT value of the modelled protein structure based on Swiss model. Four our study, we selected the best structure of the modelled protein based on the templated structure AlphaFold protein structure. For calculating the average confidence level(pLDDT) of this structure, we used the code. Here we added two code, one is based on atom based and another one is residue based calculation of pLDDT value.